Restore verbose mkpath
When running `brew -v link someFormula`, `mkdir` commands used to be
logged to the output until
https://github.com/Homebrew/legacy-homebrew/commit/f899878220668c7c7f0fcf43c6d294a52b7e79ed.

That commit removed the logging, but its stated rationale was to remove
only the counting. Hence, for symmetry with `rmdir`, it appears more
coherent to also log directory creation.

I restored the old code, modulo the counting and minor adaptation. In
particular, the output uses `mkdir` instead of `mkpath`, so that it can
be interpreted as a Unix command.