utils: avoid caching in 'paths' method
Because `paths` is a global method, the instance variable used for the
cache gets attached to whatever happens to be the receiver at the call
site. This causes the creation of multiple caches and can additionally
lead to inconsistent results if the environment is modified between
calls to `paths`.

Fix by eliminating the cache as `paths` is not performance critical and
is used in only two places throughout Homebrew:

- Homebrew.external_commands
- Homebrew::Diagnostic::Checks (by multiple methods)