diagnostic: remove MacGPG2 check
This hasn't been an issue for upstream stable releases for 3 years, since March
2013, and hasn't been an issue for pre-release versions for 4 years.

That release, and indeed the latest modern GPGTools releases, support 10.6 and
above so there's no reason to suspect people are actually encountering the old versions
in the wild with enough regularity to merit this being a permanent part of our
codebase any more.

In the last two years Homebrew has seen one Issue where MacGPG2 was the problem,
and that wasn't reproducible at the time (and still isn't), and the `doctor` check
likely wasn't even raised there. There has only been one Issue where the MacGPG2
`doctor` check was raised in that two year period.

I think it's fair to treat this as an user configuration outlier now rather than
an issue we need to be constantly on guard for.

Ref: https://github.com/Homebrew/legacy-homebrew/commit/dfb171b173e03f558aba9ef516033d75f642a2e2
Ref: https://github.com/Homebrew/legacy-homebrew/issues/12238
Ref: https://github.com/Homebrew/legacy-homebrew/commit/046498b72f2376e1755505cb924dab5cba463c45