Unify HOMEBREW_LIBRARY_PATH initialization
Stop initializing `HOMEBREW_LIBRARY_PATH` and manipulating `$LOAD_PATH`
directly in `brew.rb`. This is a leftover from when it had a shebang,
was executable, and thus needed to setup itself. Now we can safely move
this initialization into `brew.sh`.

The major benefit is that some rather awkward code in the integration
tests can be avoided, differences between test and non-test code are
further minimized, and everything is a bit less order-dependent.