gpg_requirement: drop gpg1 support
We're going to leap over Debian in renaming GPG 1.x `gpg1`. We're not going
to do it completely, you'll still be able to plonk the `libexec/bin` path of `gpg`
in your PATH to access a vanilla `gpg` 1.x executable if you want to, but we're not
going to actively keep adding gpg1 support to formulae going forwards.

There's really no reason why 99.9% of projects should not or cannot use `gpg2`
these days, and unlike a later transition to GPG 2.1.x the transition to `gpg2`
should be pretty smooth. Yes, famous last words I'm absolutely sure.