check_linkage: print full name for other taps only
`depends_on` only requires full name when the dependency is a member of
some other tap. Even though redundant full names work, they are rarely
used in practice, so to avoid confusion, it would be best for
`brew audit` to follow the use-non-full-name-when-possible convention
when it prints undeclared linkage failures.

For example, if the formula being audited is in homebrew/science, print

```
      depends_on "homebrew/dupes/ncurses" => :linked
      depends_on "htslib" => :linked
```

not

```
      depends_on "homebrew/dupes/ncurses" => :linked
      depends_on "homebrew/science/htslib" => :linked
```

Note that this change is irrelevant with respect to core formulae since
their full names are not prefixed with "homebrew/core" anyway.

Thanks to Xu Cheng for contributing the code in this commit.