BumpFormulaPR learns --remote option
When bumping formula that are in taps owned by the developers doing the
bumping, it would be convenient if the submitted PRs could be opened
from the owner repo itself. (For example, I am a contributor to nodenv
tap, and I would like to open PRs using branches from the nodenv org
rather than a fork.)

Presently, bump-formula-pr derives the remote name by running hub-fork.
This actually forks the repo on github. However, if I am already an
owner or contributor of the origin remote, then there is no need to
create a fork. This PR teaches bump-formula-pr to accept the `--remote=`
option, accepting a remote name as the argument.

If provided, hub-fork is not called. Instead, git-push is executed
against the given remote name.