Accept "Formulae" as directory of Formula files
The plural "formulae" is used frequently throughout the Brew commands and documentation.
Given that, it struck me as odd that the plural form was *not* accepted as the name for the directory of formulae, particularly given you'd generally only want the directory if it were to contain multiple formulae!

This updates it to accept `Formula`, `Formulae`, `HomebrewFormula` and `HomebrewFormulae`.