Add failing tests for PR #2466
This commit adds a couple of tests for the utility methods in
`utils/curl.rb`, including a few common and edge cases. One of the
edge cases with large stdout/stderr outputs should fail due to an
issue in the `curl_output` method.

Note that this commit also introduces a mock for the `/usr/bin/curl`
binary, which gets injected into the methods under test via the
`HOMEBREW_CURL` environment variable.

References:

- https://github.com/Homebrew/brew/pull/2466

- https://github.com/Homebrew/brew/blob/d84655efb9dba5840d75e08da5b69f5fd0c56744/Library/Homebrew/utils/curl.rb#L40