diagnostic: don't tap homebrew/core
Attempting to look up a formula when the core tap has not been set up
yet made Homebrew silently clone it, which meant that `brew doctor` took
a very long time with no indication of what was happening.

Since both of the diagnostic checks that caused this behaviour will only
fail if a homebrew/core formula is installed, it seems a little silly to
clone it just for the purpose of checking that.