Whitelist keg-only `ncurses` for `brew audit`
With regards to `brew audit`, this commit whitelists the `ncurses`
formula, which is tagged `keg_only :provided_by_macos`.

This tag had caused `brew audit --new-formula sc-im` to fail with the
following error message:

> Dependency 'ncurses' may be unnecessary as it is provided by
> macOS; try to build this formula without it.

However, the `sc-im` formula depends on the GNU ncurses library in a
recent version, just like the one Homebrew provides as a separate
formula.

Because the macOS-provided ncurses is no drop-in replacement, this
commit whitelists ncurses in `audit.rb`.

See also:

[1] https://github.com/Homebrew/homebrew-core/pull/14067#issuecomment-335046151