dev-cmd/bottle: formula rebuild `--onto` another
When your workflow is based around bottling on a branch, rather than on
master, you work with formula files from your branch. However, for
purposes of bottle rebuilds, you'd like to consider the file an
extension of a tapped formula.

The current behavior, using `--force-core-tap`, always sets rebuild to
`0`. This can result in bottle hashes falling out of sync with users, as
every build will generate the same filename.

This commit adds an `--onto` flag to `brew bottle` that allows you to
specify a core or tapped formula to consider the file an extension on
for rebuild purposes.

This allows for correct bottle filenames and build metadata, when
rebuilding from formula files.

Change-Id: I3e5864b73ccbd3c0e32aa5f045128a83ec46a1e6