Propagate quarantine if the unpacking did not
Non-dmg containers do not propagate the container's extended attributes
when unpacking. If one checks the quarantine status when moving
artifacts, these will not have the 'com.apple.quarantine' status,
rendering all the efforts futile.

With this commit, quarantine status is added upon download and checked
after the primary container is unpacked. If the quarantine attribute is
not detected in the staged path's contents, it is recovered from the
downloaded file and manually applied to the unpacked contents.