Let the user disable quarantining
As requested by @vitorgalvao, this new global flag permits to not
quarantine the Cask for the current operation.

Do notice that this won't have any effect in cached, quarantined
downloads i.e. I don't clear the attribute if it has already been set,
but I set if its previous download didn't quarantine it.