README: remove myself
I resign from Homebrew with immediate effect, and note for the record
this resignation does not come of free will.

Myself & ilovezfs had a frequently volatile relationship, but there was
one issue in particular we consistently agreed on right from the start.
In time people will join the dots, and someone will force the issue,
and tea will be spilled.

For now, that’s all I’ll say, but know from this point on:
  If a man talks shit, then I owe him nothing.
Goodbye Homebrew.