doctor: Check for /usr/include following Xcode 10
In Xcode 10, headers are no longer installed to /usr/include, which
poses a problem when software tries to be built on platforms with it
installed. This means that a fresh copy of Mojave, Homebrew, rbenv, and
ruby-build cannot build and install rubies.

Whether or not this impacts homebrew directly, it does affect software
installed with it -- which can be annoying to debug.

This thread on the Apple Developer Forums is the source for this:

-> https://forums.developer.apple.com/thread/104296