Fix missing shared symlinks for versioned Postgres
When creating a formula that depends on a versioned Postgres, e.g.
`postgres@10`, with an intention to install to shared Postgres
directory, the symlinks were not created.

`postgres` in `SHARE_PATHS` part of the `case` statement was only
matching when passed `postgres`, but not `postgres@10`.

Example formula:

    class FooExtension < Formula
      desc "Foo extension for PostgreSQL"
      homepage "https://www.example.com/"
      url "https://github.com/zombodb/zombodb/archive/v10-1.0.0.tar.gz"
      sha256 "2023180933a9513045f9a2e7b67868a8118e8b5c80038e155bba907195596307"
      depends_on "postgresql@10"

      def install
        mkdir "stage"
        system "mkdir -p stage/FAKE/share/postgresql@10/extension/*"
        system "echo fake > stage/FAKE/share/postgresql@10/extension/fake.control"
        (share/"postgresql@10/extension").install Dir["stage/**/share/postgresql@10/extension/*"]
      end
    end

That formula when installed was not creating `fake.control` in
/usr/local/share/postgresql@10/extension.

With this fix it does.