Add TMP to bin/brew whitelist to use TMP for downloading bottles
Ruby uses TMPDIR, TEMP or TMP or /tmp for its temporary directory.
Because Brew does not whitelist TMPDIR, TEMP or TMP, Ruby will
always use /tmp to download and process bottles. If /tmp is full
and you do not have control over /tmp, you cannot use bottles.
TMPDIR is used by *nix and TEMP is used by Windows. TMP does not
seem to be used by any OS so it can be used as a variable to store
an alternative temporary directory without impacting the OS.