Fail: preserving .DS_Store-cleaning per-formula.
So, I gave this an honest shot here, but after 4 hours I just haven't been able to implement what you suggested @MikeMcquaid, that is, being able to clean .DS_Store files when called for with individual formulae. This was my last go at it, but I'm afraid I've gotta move on at this point. :(