Make vendor-install.sh script more robust
`vendor-install.sh` script relies on the presence of the
'$VENDOR_DIR/portable-$VENDOR_NAME-version' file.
In situations when user (accidentally) deletes this file or
the directory containing it, executing the `brew` command will fail with:

    /usr/local/Homebrew/Library/Homebrew/cmd/vendor-install.sh: line 239: /usr/local/Homebrew/Library/Homebrew/vendor/portable-ruby-version: No such file or directory
    ==> Downloading https://homebrew.bintray.com/bottles-portable-ruby/portable-ruby--2.6.3.mavericks.bottle.tar.gz
    Already downloaded: /Users/mbelkin/Library/Caches/Homebrew/portable-ruby--2.6.3.mavericks.bottle.tar.gz
    ==> Pouring portable-ruby--2.6.3.mavericks.bottle.tar.gz
    Error: Failed to vendor ruby .
    Error: Failed to install vendor Ruby.

This fixes situations like these by fetching the official Homebrew/brew
tap and completely restoring the directory containing this file.