Keg: apply ad-hoc signature on 10.15 or later
This is an alternative to #9040. In that case, we rely on ruby-macho to
apply code signatures after changing ID or install_name. Here, we instead
do that ourselves so that we can choose what OS to apply it on. Like in
that PR, we swallow any failures so that we don't inadvertently break
more exotic packages like OpenJDK (which caused issues in #8922).