bin/brew: re-exec self under Rosetta if needed
As discussed in #9177, we've reserved /usr/local as
the prefix to use on Apple Silicon Macs for an Intel
installation of Homebrew to be run under Rosetta.

If a user has a /usr/local installation on an Apple
Silicon Mac, and executes /usr/local/bin/brew, we
should make sure we're running that copy of brew as
Intel under emulation in order to avoid confusion.
This makes that change.