bin/brew: Enforce use of MacOS versions of tools.
brew assumes that the MacOS (BSD-style) versions of dirname/readlink are used by the user. This may not be the case if the user chooses to use the Homebrew installed (GNU-style) coreutils instead. I.e. use `PATH="/usr/local/opt/coreutils/libexec/gnubin:$PATH"`, as proposed after installing coreutils.
This patch make sure that the assumed version of the tools is always used.