Add more reliable way to run homebrew-pypi-poet
Replace the `virtualenvwrapper` method with one based on the `pipenv`
formula, a dependency manager but doubles as a `virtualenvwrapper`
alternative.

A non-brewed installation of virtualenvwrapper is not always reliable.

For me, it fails like this:

```
$ source /Users/claudia/Library/Python/3.9/bin/virtualenvwrapper.sh
/usr/bin/python: No module named virtualenvwrapper
virtualenvwrapper.sh: There was a problem running the initialization hooks.
If Python could not import the module virtualenvwrapper.hook_loader,
check that virtualenvwrapper has been installed for
VIRTUALENVWRAPPER_PYTHON=/usr/bin/python and that PATH is
set properly.
```