Add additional info for updating Command Line Tools
Add an additional note pointing users to a wiki page showing which Command Line Tool version is the latest available for their OS.

This is useful because the `softwareupdate` or `xcode-select --install` commands don’t always  manage to install the latest CLT, and a manual download and install is sometimes necessary. With a manual download, there’s no easy way to know which CLT download is the latest available & compatible one for a particular OS, which is what makes the wiki reference page very useful.