download_strategy: clear referer by default with curl
This fixes the situaiton when a user `.curlrc` contains e.g. `--referer ';auto'`
which in turn may break download URLs unless reset to the default (empty)
value. Such download URLs are anything under `https://downloads.sourceforge.net/`.

Also switch to the long-form of the -e option, which is --referer, for readability.