formula_cellar_checks: exclude llvm from Python linkage check
The LLDB Python bindings embed a Python interpreter into LLDB, so
explicit linkage to a Python interpreter is needed. [1]

This change will allow us to build the Python bindings for LLDB, which
improves feature-parity with system LLDB. Currently, these bindings are
not built because they fail the `check_python_framework_links` audit.

[1] https://blog.tim-smith.us/2015/09/python-extension-modules-os-x/