fix(software_spec): use -- for fetch URL that we use for the bottle
The other methods on this class all use the `--`, which causes bottles
to be published with `--` but attempted to be downloaded with `-` (which
errors).

I believe this doesn't affect homebrew-core's own bottles any more
because the new GitHub Packages scheme doesn't use the filename in the
URL at all. However it seems to break other bottle `base_url`s.