utils/pypi respect ignore_non_pypi_packages
elsewhere on this file, the ignreo_non_pypi_packages flag is used
to continue to make progress despite the existence of non-pypi resources

There is one check in the update_python_resources function that
will always odie if there is a non-pypi resource. This prevents
brew bump-forumla-pr from being run on formulas that contain
both pypi and non-pypi resources.

This PR modifies said check to not call odie when such resources
are encounted