cpu: add AVX-512 and NEON helper functions
Add helper functions for AVX-512 and NEON extensions.

Note that AVX-512 is not a monolithic extension but rather a family
of extensions, of which vendors can choose to implement some or all
of them. We expose all of the individual extensions here so that each
can be checked individually. "avx512f" is the "foundation" extension
that is required for any other AVX-512 functionality, but it seems
inaccurate to just have a single function "avx512?" that checks only
this one extension. Similarly, it seems inaccurate to just have a
single function "avx512?" that checks all of the extensions (in fact,
AFAIK no single CPU currently implements all of the extensions).

NEON aka Advanced SIMD is an ARM extension analagous to Intel's
SSE. Note that on Linux, the feature flag can be either "neon" or
"asimd".