cmd/uses: add `--skip-recursive-build-dependents` flag
We use `brew uses` to determine the list of dependent formulae to test
in `test-bot`. By default, we generate this list with the invocation

    brew uses --formula --include-build --include-test --recursive <formula>

This results in our testing formulae that do not need to be tested. As
an example, suppose `foo` is a build dependency of `bar`, and `baz`
depends on `bar`. Since we don't rebuild `bar` when testing `baz`, then
no changes in `foo` are relevant to `baz`, and any test result from
testing `baz` as a dependent of `foo` is not meaningful.

One instance where this is particularly wasteful is in `go` PRs. `envoy`
and `envoy@1.18` both declare a build dependency on `bazelisk`, which in
turn declares a build dependency on `go`. The two formulae also take
about four hours each to build.

Let's cut the CI time we spend testing formulae that don't need to be
tested by adding a `--skip-recursive-dependents` flag to `brew uses`.
Adding this flag will omit all dependents of a build dependent from the
output.

For reference, here is the list of formulae that are present without the
`--skip-recursive-build-depenents` flag but are no longer there with it:

arduino-cli
bat-extras
crun
envoy
envoy@1.18
kube-ps1
kubectx
kubernetes-cli
kubie
modd
newrelic-infra-agent
terraform@0.11