Provide a way for regular users to ignore xcode check
In some corporations folks run vendored versions of xcode where updating
to the latest version of xcode is simply not possible with the toolchain
they are working with. Given there is already a way to bypass the xcode
check in github actions, there should be a reasonable way for most users
to bypass the check too if they would like knowing the risks it may
entail.

There is already a rather widely shared solution similar to this:
https://stackoverflow.com/a/56252549

However this is brittle and is prone to reverting if a brew update is
run (which the default brew configuration will do when installing a new
package https://github.com/Homebrew/brew/issues/9285)