keg_relocate: fall back to formula when tab is missing
Installing Java-dependent formulae from third-party taps and bottle
mirrors can, for a variety of reasons, result in the tab going missing.
See Homebrew/discussions#2530.

One result of this is that kegs are not relocated correctly, and this
produces confusing errors for users at runtime. Let's fix that by
falling back to the formula for dependency information when we need to.