keg: allow instantiating kegs not rooted in the cellar
I often download bottles directly from a GitHub workflow and unpack them
in a temporary directory. The `Keg` class provides a number of
convenience methods for examining the contents of unpacked bottles (e.g.
`Keg#mach_o_files`), so I often find myself temporarily commenting out
the check for a keg being rooted in the cellar.

It would simplify things a lot for me if I could just pass an extra
argument to `Keg.new` to skip checking for whether the `Keg` I'm trying
to instantiate is rooted in the cellar.