shims: make CC shim callable without a symlink
Currently, the CC shim is only callable via its symlink in shims/mac/super:
https://github.com/Homebrew/brew/blob/master/Library/Homebrew/shims/mac/super/cc

In particular, it expects xcrun to be in the same directory:
https://github.com/Homebrew/brew/blob/master/Library/Homebrew/shims/super/cc#L473

This poses a problem for build systems such as Bazel that resolve symlinks during toolchain configuration.
Bazel seems to have a good reason for resolving symlinks, which is a bug in clang related to symlink handling:
https://github.com/bazelbuild/bazel/blob/a987b98ea0d6da2656c4115568ef9cbe8a164550/tools/cpp/unix_cc_configure.bzl#L312

Therefore, it seems preferable to make CC more-flexible in how it is called.

I do this by checking both possible locations for xcrun.

This change should enable using Bazel C++ rules in Brew, which will allow us to remove the shim bypass from Bazel
and enable new formulas:
https://github.com/Homebrew/homebrew-core/blob/master/Formula/bazel.rb#L48