#curl_download: Use first response headers
Before Homebrew/brew#11252 was merged, #curl_download used the headers
from the first response to check for `accept-ranges` and
`content-length` headers. This behavior was incorrect and it should
have been checking the last response headers, as earlier responses
would be redirections that may omit the `accept-ranges` header and/or
use a different `content-length` value than the final response.

I intended to maintain the "first" behavior in #11252 and switch to
`#last` in a follow-up PR but I accidentally incorporated this change
into the aforementioned PR. This commit should not be merged and it's
simply for the sake of explaining this change after-the-fact.