workflows: add nightly linux tests
Test a simple simple brew installation and "brew install hello" step for various distributions.

This makes sure everything is working well.
We use our installer as we want to test everything end to end.

The idea is also to be able to tell users how to install curl/git/tar on older linux versions when needed.

In the future, these images might be uploaded to our docker registry, and I plan to use them
to test migrations (from gcc5 to gcc7, and potentially for building binutils on ubuntu 18.04).