brew.sh: create `current` symlink in `pkgconfig` directory
`pkg-config` currently hardcodes the OS version it was built on in its
pc path in order to find the `.pc` files shipped with Homebrew. This
requires reinstallation of `pkg-config` with OS upgrades, which is
something we currently do not handle.

We can avoid this by automatically creating a `current` symlink that
points to the right directory whenever `brew.sh` runs. This will require
a rebuild of `pkg-config` that points it to the `current` directory.

In order to simplify identifying the right directory to symlink to, I've
dropped the patch version from `HOMEBREW_OS_VERSION` on Catalina and
earlier. I don't think this should break anything, but I can try a
different approach if this will cause problems.